refactoring_method
