+++ src/main/java/ru/mystamps/web/service/MailService.java
@@ -23,5 +23,4 @@ import ru.mystamps.web.service.dto.SendUsersActivationDto;
 public interface MailService {
 	void sendActivationKeyToUser(SendUsersActivationDto activation);
 	void sendDailyStatisticsToAdmin(AdminDailyReport report);
-	String prepareDailyStatistics(AdminDailyReport report);
 }
