+++ src/main/java/ru/mystamps/web/service/ReportService.java
@@ -15,34 +15,13 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package ru.mystamps.web.controller;
+package ru.mystamps.web.service;
 
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import lombok.RequiredArgsConstructor;
-
-import ru.mystamps.web.Url;
-import ru.mystamps.web.service.CronService;
-import ru.mystamps.web.service.ReportService;
+import ru.mystamps.web.service.dto.AdminDailyReport;
 
 /**
- * @author Maxim Shestakov
+ *@author Maxim Shestakov
  */
-@Controller
-@RequiredArgsConstructor
-public class ReportController {
-
-	private final ReportService reportService;
-	private final CronService cronService;
-
-	@GetMapping(path = Url.DAILY_STATISTICS, produces = "text/plain; charset=UTF-8")
-	@ResponseBody
-	public String showDailyReport() {
-		return reportService.prepareDailyStatistics(
-			cronService.getDailyReport()
-		);
-	}
-
+public interface ReportService {
+	String prepareDailyStatistics(AdminDailyReport report);
 }
